update to simpler coding until further understanding of randomizing with a button click is understood
Code was not functional due to lack of understanding of randomizing an outcome after clicking a button, instead a simple button click should execute a font change of the header for the first button from a default font to a different font. Second button can toggle an image to hide and reveal itself.